Shell Batch Fix Part 1
If you were expriencing issues building Choacury from source, this is a minor fix for that, well more accurately the first half of the fix.

The second half is hopefully coming soon and not everything has been fixed! You should see far less errors but it may not compile properly because a commit has screwed up the shell. This is me trying this fix this damage.

If you want to  make this easier feel free to commit some fixes to the shell too!